requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
DateTime==4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
imgurpython
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
line-bot-sdk==1.8.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
pandas-datareader==0.9.0
Pillow==7.2.0
psycopg2==2.8.5
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
twder==0.1.4
urllib3==1.25.10
Werkzeug==1.0.1
zope.interface==5.1.0
